+++ app/lib/displayOpenResults.js
@@ -1,0 +1,5 @@
+function displayOpenResults(req) {
+  return req.query.open ? req.query.open.toLowerCase() === 'true' : false;
+}
+
+module.exports = displayOpenResults;