+++ app/lib/utils.js
@@ -1,0 +1,7 @@
+function deepClone(obj) {
+  return JSON.parse(JSON.stringify(obj));
+}
+
+module.exports = {
+  deepClone,
+};